[Refactor] main에서 scroll 타입 분리
