+++ src/types/main/index.d.ts
@@ -23,15 +23,3 @@ export interface FloatingButtonProps {
     onScrollToStart: () => void;
     activeSection: string;
 }
-
-export interface SectionRefs {
-    [key: string]: React.RefObject<HTMLDivElement>;
-}
-
-export interface ScrollContextType {
-    containerRef: React.RefObject<HTMLDivElement>;
-    sectionRefs: SectionRefs;
-    scrollToRef: (sectionKey: keyof SectionRefs) => void;
-    activeSection: string;
-    setActiveSection: (activeSection: string) => void;
-}
